more practice

+++ Lab02_Functions/Practice20240916/01_SignOfIntegerNumbers/01_CenterPoint.cpp
@@ -1,21 +1,21 @@
 #include <iostream>
 
-double distFromZero(int a, int b) {
+double distFromZero(double a, double b) {
     return (a * a + b * b);
 };
 
-void print(int a, int b) {
+void print(double a, double b) {
     std::cout << '(' << a << ", " << b << ')';
 }
 
-void compare(int a1, int b1, int a2, int b2) {
+void compare(double a1, double b1, double a2, double b2) {
     (distFromZero(a1, b1) > distFromZero(a2, b2)) ? print(a2, b2) : print(a1, b1);
 }
 
 
 int main()
 {
-    int x1, y1, x2, y2;
+    double x1, y1, x2, y2;
     std::cin >> x1 >> y1 >> x2 >> y2;
 
     compare(x1, y1, x2, y2);
